Crate filecoin_proofs_api

Proofs library for Filecoin open blockchain network.

This goal of the filecoin-proofs-api crate is to provide the proving and verification mechanisms used within Filecoin, to ensure Storage Providers are properly storing client data, and retaining that data over time. The ‘Proofs’ are used to assert that the work was done properly, and the nodes on the network are able to verify proofs to maintain trust across the distributed storage network. The proving system used by Filecoin is based on Groth16. Specific poofs include Proof-of-Spacetime, and Proof-of-Replication.

For further information, please see the Filecoin specification

Re-exports

pub use fr32;

Modules

post

Generation and verification of Proof-of-Spacetime Merkle tree and zk-SNARK proofs.

seal

Proof-of-Replication for sealing, unsealing, and verifying data sectors

update

Update data within existing sealed sectors.

Structs

FallbackPoStSectorProof

PaddedBytesAmount

PartitionProofBytes

PartitionSnarkProof

PieceInfo

PrivateReplicaInfo

PublicReplicaInfo

SectorId

Identifier for a single sector.

UnpaddedByteIndex

UnpaddedBytesAmount

Enums

ApiFeature

The ApiFeature enum is used for optional features that the network can use and recognize, but in no way is required to be used.

ApiVersion

The ApiVersion enum is used for mandatory changes that the network must use and recognize.

PoStType

RegisteredAggregationProof

Available aggregation of zk-SNARK proofs.

RegisteredPoStProof

Available PoSt proofs.

RegisteredSealProof

Available seal proofs.

RegisteredUpdateProof

Available proofs for updating sectors

StorageProofsError

Custom error types

Constants

NODE_SIZE

Traits

MerkleTreeTrait

Trait used to abstract over the way Merkle Trees are constructed and stored.

Functions

get_parameter_data

Get the correct parameter data for a given cache id.

get_verifying_key_data

Get the correct verifying key data for a given cache id.

Type Aliases

AggregateSnarkProof

ChallengeSeed

Commitment

OrderedSectorSet

An ordered set of SectorIds.

ProverId

SnarkProof

Ticket

VanillaProof