Module filecoin_proofs_api::post

Generation and verification of Proof-of-Spacetime Merkle tree and zk-SNARK proofs.

Functions

generate_fallback_sector_challenges

Generates the challenges per SectorId required for either a Window Proof-of-Spacetime or a Winning Proof-of-Spacetime.

generate_single_vanilla_proof

Generates a single vanilla proof required for either Window Proof-of-Spacetime or Winning Proof-of-Spacetime.

generate_single_window_post_with_vanilla

Generates a Window Proof-of-Spacetime with provided vanilla proofs of a single partition.

generate_window_post

Generates a Window Proof-of-Spacetime.

generate_window_post_with_vanilla

Generates a Window Proof-of-Spacetime with provided vanilla Merkle tree proofs.

generate_winning_post

Generates a Winning Proof-of-Spacetime.

generate_winning_post_sector_challenge

Given some randomness and the length of available sectors, generates a sector challenge.

generate_winning_post_with_vanilla

Generates a Winning Proof-of-Spacetime with provided vanilla Merkle tree proofs.

get_num_partition_for_fallback_post

For window PoSt only, gets the number of partitions based on proof type.

merge_window_post_partition_proofs

Merge multiple window PoSt partition proofs into a single zk-SNARK proof.

verify_window_post

Verifies a window Proof-of-Spacetime.

verify_winning_post

Verifies a winning Proof-of-Spacetime.