Function filecoin_proofs_api::seal::aggregate_seal_commit_proofs

pub fn aggregate_seal_commit_proofs(
    registered_proof: RegisteredSealProof,
    registered_aggregation: RegisteredAggregationProof,
    comm_rs: &[Commitment],
    seeds: &[Ticket],
    commit_outputs: &[SealCommitPhase2Output]
) -> Result<AggregateSnarkProof>

Given a porep_config and a list of seal commit outputs, this method aggregates those proofs (naively padding the count if necessary up to a power of 2) and returns the aggregate proof bytes.

Arguments

registered_proof - Selected seal operation.

registered_aggregation - Aggregation proof types.

seeds - Ordered list of seeds used to derive the PoRep challenges.

commit_outputs - Ordered list of seal proof outputs returned from seal_commit_phase2.

Returns aggregate of zk-SNARK proofs in AggregateSnarkProof.