Function filecoin_proofs_api::seal::seal_commit_phase2_circuit_proofs

pub fn seal_commit_phase2_circuit_proofs(
    phase1_output: SealCommitPhase1Output,
    sector_id: SectorId
) -> Result<SealCommitPhase2Output>

Generates zk-SNARK proof for sector replica. Must be called with output of seal_commit_phase1.

This variant of seal_commit_phase2 is intended specifically for returning the circuit proofs of a NonInteractivePoRep proof, such that it can later be aggregated with other NonInteractivePoRep proofs. Using it outside of this context may cause errors.

Arguments

phase1_output - Struct returned from seal_commit_phase1 containing Merkle tree proof.

sector_id - ID of the sector, usually relative to the miner.

Returns SealCommitPhase2Output struct containing vector of zk-SNARK proofs.