Module filecoin_proofs_api::update

Update data within existing sealed sectors.

Functions

empty_sector_update_decode_from

Reverses the encoding process and outputs the data into out_data_path.

empty_sector_update_decode_from_range

Reverses the encoding process for a certain range.

empty_sector_update_encode_into

Encodes data into an existing replica.

empty_sector_update_remove_encoded_data

Removes encoded data and outputs the sector key.

generate_empty_sector_update_proof

Generate updated proof from an empty sector replica.

generate_empty_sector_update_proof_with_vanilla

Generate updated proof from an empty sector provided the vanilla proof and new commitment.

generate_partition_proofs

Generate all vanilla partition proofs across all partitions.

verify_empty_sector_update_proof

Verify an empty sector update proof provided the proof and new and old commitments.

verify_partition_proofs

Verify all vanilla partition proofs across all partitions.