Function filecoin_proofs_api::seal::seal_commit_phase1

pub fn seal_commit_phase1<T: AsRef<Path>>(
    cache_path: T,
    replica_path: T,
    prover_id: ProverId,
    sector_id: SectorId,
    ticket: Ticket,
    seed: Ticket,
    pre_commit: SealPreCommitPhase2Output,
    piece_infos: &[PieceInfo]
) -> Result<SealCommitPhase1Output>

Generate Merkle tree proofs for sector replica. Must be called with output from seal_pre_commit_phase2.

Arguments

cache_path - Directory path to use for generation of Merkle tree on disk.

replica_path - out_path from seal_pre_commit_phase2, which points to generated sector replica.

prover_id - Unique ID of the storage provider.

sector_id - ID of the sector, usually relative to the miner.

ticket - The ticket used to generate this sector’s replica-id.

seed - Interactive randomnessthe seed used to derive the Proof-of-Replication (PoRep) challenges.

piece_infos - The piece info (commitment and byte length) for each piece in the sector.

Returns vanilla Merkle tree proof for use by seal_commit_phase2.