Function filecoin_proofs_api::update::empty_sector_update_decode_from

pub fn empty_sector_update_decode_from<R, S, T, U>(
    registered_proof: RegisteredUpdateProof,
    out_data_path: R,
    replica_path: S,
    sector_key_path: T,
    sector_key_cache_path: U,
    comm_d_new: Commitment
) -> Result<()>
where
    R: AsRef<Path>,
    S: AsRef<Path>,
    T: AsRef<Path>,
    U: AsRef<Path>,

Reverses the encoding process and outputs the data into out_data_path.

Arguments

registered_proof - Selected sector update proof.

out_data_path - File path to output decoded data.

replica_path - File path of replica.

sector_key_path - Path to sector key originally used to seal sector.

staged_data_path - Path to staged data to encode into existing replica.

comm_d_new - Data commitment from updated replica.